Added a bunch of sorting algorithms, heap sort currently not working

+++ Revised 2/Merge Sort Algorithm.vb
@@ -1,15 +1,15 @@
 ﻿Module Merge_Sort
-    Function mergesort(ByVal m As List(Of Integer))
+    Function mergesort(ByVal m As List(Of Double))
         If m.Count <= 1 Then Return m
-        Dim left As List(Of Integer) = m.GetRange(0, m.Count \ 2)
-        Dim right As List(Of Integer) = m.GetRange(m.Count \ 2, m.Count \ 2)
+        Dim left As List(Of Double) = m.GetRange(0, m.Count \ 2)
+        Dim right As List(Of Double) = m.GetRange(m.Count \ 2, m.Count \ 2)
         left = mergesort(left)
         right = mergesort(right)
         Return merge(left, right)
     End Function
 
-    Function merge(ByVal left As List(Of Integer), ByVal right As List(Of Integer))
-        Dim result As New List(Of Integer)
+    Function merge(ByVal left As List(Of Double), ByVal right As List(Of Double))
+        Dim result As New List(Of Double)
         While left.Count > 0 And right.Count > 0
             If left.First <= right.First Then
                 result.Add(left.First)
